Add new option --insist-isnt
If we have a case where it can be seen that a constructor is not
returned, such as

    case x of
      ... -> LT
      ... -> GT

we add

    insist (x `isn't` EQ)

Could make it so that the axiom is only laid down when the thunk is
delayed. Because when it is forced, a corresponding axiom will
be laid down anyway.

This helps for instance heap sort, but slows down some examples, like
merge sort.